+++ TP Especial/ServicioMonopatin/src/main/java/com/example/demo/model/Viaje.java
@@ -17,7 +17,7 @@ public class Viaje {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idViaje;
+	private Long idViaje;
 	
 	@Column
 	private LocalDate fechaYHoraInicio;
